feat: (WIP) Migrate test scripts to /tests folder
