+++ tests/test_audit_logs.py
@@ -7,14 +7,9 @@ from main import app
 
 client = TestClient(app)
 
-def test_read_main():
-    response = client.get("/")
-    assert response.status_code == 200
-    assert response.json() == {"Hello": "World"}
-
 SAMPLE_LOG = {
     "tenant_id": str(uuid.uuid4()),
-    "user_id":   str(uuid.uuid4()),
+    "user_id": str(uuid.uuid4()),
     "session_id": "sess-test-001",
     "ip_address": "127.0.0.1",
     "user_agent": "test-client/1.0",
@@ -25,11 +20,6 @@ SAMPLE_LOG = {
     "before_state": {"foo": "bar"},
     "after_state": {"foo": "baz"},
     "metadata": {"test_meta": True}
-}
-
-SAMPLE_TENANT = {
-    "name": "tenant-01",
-    "status": "active"
 }
 
 def test_search_log_empty():
@@ -65,7 +55,6 @@ def test_get_log_stats():
     resp2 = client.get("/api/v1/logs/stats")
     assert resp2.status_code == 200
 
-
 def test_create_new_log():
     # Create a new log
     resp = client.post("/api/v1/logs/", json=SAMPLE_LOG)
@@ -96,28 +85,6 @@ def test_create_bulk(size):
         body = resp.json()
         assert "Data inserted" in body
         assert body["Data inserted"] == size
-
-# Tenant related tests
-def test_search_tenant_after_create():
-    # Create a new record
-    resp = client.post("/api/v1/tenants/", json=SAMPLE_TENANT)
-    assert resp.status_code == 201, resp.text
-    created = resp.json()
-    log_id = created["id"]
-
-    # Search for newly created log
-    resp2 = client.get("/api/v1/tenants/")
-    data = resp2.json()["data"]
-    assert (item["id"] == log_id for item in data)
-
-def test_create_tenant():
-    resp = client.post("/api/v1/tenants/", json=SAMPLE_TENANT)
-    assert resp.status_code == 201, resp.text
-    created = resp.json()
-
-    # Assert id and created_at is generated
-    assert "id" in created
-    assert "created_at" in created
 
 def test_websocket_log_stream():
     tenant_id = uuid4()
